+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Minimum required Gradle bumped from `5.2` to `6.1` because we need `BuildService`.
 
 ## [2.2.0] - 2021-05-11
 ### Added